requirements.txt
langchain==0.3.26
langchain-openai==0.3.28
langchain-community==0.3.27
langchain-core==0.3.70
langchain-text-splitters==0.3.8
langsmith==0.4.8
langchain_community==0.3.27
langchain_openai==0.3.28
requests==2.32.4
beautifulsoup4==4.13.4
python-dotenv==1.1.1
openai==1.97.1
tiktoken==0.9.0
typing-extensions==4.14.1